# HOTT-4464 Fixing timing issues in JS specs