--- a/spec/features/search_banner_spec.rb
+++ b/spec/features/search_banner_spec.rb
@@ -13,21 +13,16 @@ RSpec.describe 'Search banner', js: true do
         page.find('.tariff-search-banner .autocomplete__input').click
 
         page.find('.tariff-search-banner .autocomplete__input').set('gold')
-        sleep 1.5
 
+        expect(page).to have_css('.tariff-search-banner .autocomplete__option')
         expect(page.find('.tariff-search-banner .autocomplete__option:first-of-type').text).to eq('gold')
 
-        using_wait_time 1 do
-          expect(page.find_all('.tariff-search-banner .autocomplete__option').length).to be > 1
-        end
+        expect(page.find_all('.tariff-search-banner .autocomplete__option').length).to be > 1
 
         expect(page.find('.tariff-search-banner .autocomplete__option:first-of-type').text).to eq('gold')
         expect(page).to have_content('goldsmiths')
 
         page.find('.tariff-search-banner .autocomplete__option:first-of-type').click
-
-        # trying to see if redirect done by JS needs some sleep to be caught up
-        sleep 1
 
         expect(page).to have_content('Search results for ‘gold’')
       end
@@ -45,16 +40,12 @@ RSpec.describe 'Search banner', js: true do
 
         page.find('.tariff-search-banner .autocomplete__input').set('dsauidoasuiodsa')
 
-        sleep 1.5
+        expect(page).to have_css('.tariff-search-banner .autocomplete__option')
 
         expect(page.find_all('.tariff-search-banner .autocomplete__option').length).to eq(1)
         expect(page.find('.tariff-search-banner .autocomplete__option:first-of-type').text).to eq('dsauidoasuiodsa')
-        sleep 1
 
         page.find('.tariff-search-banner .autocomplete__option:first-of-type').click
-
-        # trying to see if redirect done by JS needs some sleep to be caught up
-        sleep 1
 
         expect(page).to have_content('Search results for ‘dsauidoasuiodsa’')
         expect(page).to have_content('There are no results matching your query.')