### Jira link

HOTT-4464

### What?

I have added/removed/altered:

- [x] Increased the max time out for any one test in js specs from 2 to 5 seconds
- [x] Removed override code (`using_wait_until`) which dropped it from 2 to 1 seconds ??
- [x] Removed sleeps, these are replaced with calls to `has_css` - which, like `have_content`, will block until either the document element becomes available or max_wait_timeout is reached
- [x] Changed the cuprite config to wait longer for chrome to become available

### Why?

I am doing this because:

- The timing issues shouldn't require sleeps in specs to resolve - allowing Capybara to wait longer for a `have_css`/`have_content` condition to become true should resolve this whilst also allowing it to complete faster if it becomes true sooner
- The drops of wait timer were likely to increase failures rather than reduce them
- The cuprite config change is a bit speculative, its referenced in this github ticket against capybara - https://github.com/rubycdp/cuprite/issues/127. It shouldn't take longer as I understand it

### Deployment risks (optional)

- Low, should hopefully improve test suite reliability